chore: seperate util fn that uses storage module
